pyproject.toml
[project]
name = "horeb"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "typer>=0.12.0",
    "pydantic>=2.0.0",
    "pythonbible>=0.15.0",
    "json-repair>=0.28.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
horeb = "horeb.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests as live API calls (deselect with '-m not integration')",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
]
